+++ ktlint/src/main/kotlin/com/pinterest/ktlint/internal/GenerateEditorConfigSubCommand.kt
@@ -31,9 +31,9 @@ class GenerateEditorConfigSubCommand : Runnable {
                 text = "",
                 ruleSets = ktlintCommand.rulesets
                     .loadRulesets(
-                        ktlintCommand.experimental,
                         ktlintCommand.debug,
-                        ktlintCommand.disabledRules
+                        ktlintCommand.disabledRules,
+                        ktlintCommand.experimentalRules
                     )
                     .map { it.value.get() },
                 userData = mapOf(